+++ nyx/moirai_ecs/system/base_systems.py
@@ -6,9 +6,17 @@ from.
 """
 
 from abc import ABC
+from typing import TYPE_CHECKING
 
-from nyx.moirai_ecs.entity.moirai_entity_manager import MoiraiEntityManager
+if TYPE_CHECKING:
+    from nyx.nyx_engine.nyx_engine import NyxEngine
 
 
 class BaseSystem(ABC):
-    """The base system class, which all systems in the ECS architechture inherit from."""
+    """The base system class, which all systems in the ECS architecture inherit from."""
+
+    @property
+    def engine(self) -> "NyxEngine":
+        """Return the engine instance."""
+        from nyx.nyx_engine.nyx_engine import NyxEngine
+        return NyxEngine()